Fix bug where Confirm Destroy and Reorder pages don't render for Non Whitehall Link members
Document Collections Confirm Destroy and Reorder pages for Group Members were breaking due to the fact that it was expecting a document membership and loading the latest edition. When in the case of being added by URL, the structure is different and title needs to be loaded differently.

Fixed by using the helper that has been created for this purpose.

Co-Authored-By: Mark Taylor <mark.taylor1@digital.cabinet-office.gov.uk>
Co-Authored-By: Farah Mehmood <farah.mehmood@digital.cabinet-office.gov.uk>