Do not destroy default news image on organisation destroy
Dependent news articles might still be using the default news images.

We looked into trying to republish news articles when organisation is destroyed, by using
an `after_destroy` callback. This proved to be difficult since the organisation is gone
and we cannot link back the news articles to the organisation.

Co-authored-by: Tuomas Nylund <tuomas.nylund@digital.cabinet-office.gov.uk>
Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>