Remove awareness of what needs to be published from the CreateAssetWorker
Documents need to be republished and dependent document need to be republished
as well when all asset variants finish uploading to asset manager.

CreateAssetWorker doesn't really need to know what should be republished and what not.
That information belongs to the business model classes.

This refactoring moves the responsibilities to more appropriate places.

Co-authored-by: syed-ali-tw <syed.ali@digital.cabinet-office.gov.uk>
Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>
Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>