Extract test setup
I considered using `let()` blocks instead of `setup()` for this, but I
ended up confusing myself because the `create()` calls need to happen
before the code under test executes, which made the tests quite awkward
to write.

We would need to call expected_hash _before_ calling the presenter, to
make sure the let blocks had created all the organisations in the
database. It's much clearer what's happening if we use setup instead.