Always present /government/organisations in English
We recently encountered a bug where the /government/organisations page
was appearing in Welsh, with many organisations missing.

This was because a Welsh translation of an organisation had been
published, which triggered its after_save callback -
republish_organisations_index_page_to_publishing_api

In this situation, the pseudo-global I18n.locale is set to :cy, because
we're in the process of publishing a Welsh translation of an
organisation. However, the /government/organisations page should only
ever be presented in English.

We used to have a separate /government/organisations.cy path for the
Welsh version of this page, but now OrganisationsIndexPresenter
hardcodes the base_path. Because items in content store have to be
unique by base_path, publishing a Welsh edition overwrites the English
edition.

Additionally, OrganisationsIndexPresenter determines which organisations
to list using the listable scope. This was updated to only list
organisations with translations for the current locale in ec3979c130, to
avoid double counting Organisations which have both English and Welsh
translations. This works well if the locale is English, but misses most
organisations when the locale is set to Welsh. I've set it to filter to
English translations instead, which I think is what we want in every
situation that listable is called.

Since the Welsh version of this page has never worked properly, we
should restrict it to English editions only. We may decide to revisit
this decision later - if so, we should make sure that all the
organisations are listed in the Welsh version, even if they haven't got
Welsh translations. Just because an Organisation hasn't published a
Welsh translation of its home page, doesn't mean it doesn't exist.