Add controller spec to ensure lead image markdown isn't in body
This does 2 things:

1. updates the vaidation that checks for the markdown for a lead image
not being in the body to use the edition_lead_image rather than the
lead_image association. The lead image association returns nil until
persisted. We need to make sure the built edition_lead_image is valid.
2. adds a controller spec to ensure that update endpoint fails with
a relevant error message when a user tries to make an image the lead
image with it's markdown in the body.