Enable TopicalEventFeaturingImageData to use assets flow
Allowing TopicalEventFeaturing to implement new assets flow by whitelisting the
TopicalEventFeaturingImageData.

There is a call to publishing api being enqueued from the `AssetManagerCreateAssetWorker`
which ensures that updates are being made after each asset creation but only when all
have become available.

The filtering and the enqueueing of updates work together so that: we do not send broken
urls to publishing api if assets are not yet available; we enqueue an update from the
asset creation worker so that the healthy urls are sent when assets are available.