Filter out images with missing assets for Person when publishing to publishing api
We do not want to send broken urls to publishing api if the person has been updated
and its image assets are not ready yet, so we filter out any "unhealthy" images.

In order to ensure the correct assets are sent along once they become available,
we are republishing the person when the `CreateAssetWorker` finishes creating the
last required asset. This functionality is now enqueued from the worker and managed via the
`FeaturedImageData`. The `FeaturedImageData` acts as an intermediary between the
worker/assets and the featurable class's (e.g. Person) republishing needs.

Person has some associations that also send along the person image
so they too need to be republished when Person image changes: `Speech` and
`HistoricalAccount`. The republishing will be dealt with in a future commit,
but similarly they need to filter our unhealthy images.