Republish speeches & historical accounts if people image changes
These associations send person images to the publishing api. We have filtered any
unhealthy person images from their respective presenters so that if assets are missing
we do not send any broken urls.

With the filtering in place, we need to ensure that the correct load is sent to
publishing api, which is why we republish person and its associations.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>