Refactor assetable republishing tests
Moving tests from the worker to their respective assetable classes.

The worker is responsible for enqueueing the republishing but is otherwise agnostic as to
what should be (re)published. Whilst having the tests here provides some integration value,
they are more descriptive elsewhere.

Since the introduction of the `FeaturedImageData` for most classes that use the
`PublishesToPublishingApi` concern, this intermediary data class helps decide which featurable
to republish. The implementation logic of the actual republishing still sits in the models and the
publishing concern.

Grouping the tests for all featurable under the featured image data tests provides better
clarity on how this class orchestrates the republishing.