Add summary card list component
This is close to a copy paste of the summary card component, except that
it takes a single value for each row, in place of a key value pair.

I've had to retain using a descriptive list, rather that an ordered list
for accessibility reasons.

When using an ol or ul, li tags parent must be the ol or ul. We need
to add a a div tag to inserts rows into the component which was causing
this to not be the case.