Test edition attachments summary card in cucumber features
We want to ensure that any issues with the summary card version of edition attachments are picked up in CI. Unfortunately the scenario outline feature of cucumber doesn't allow sharing examples between scenarios, so this is quite painful. However, we thought this was still a better option than adding another cucumber task to the CI pipeline and using tags to run just the relevant tests, because having to change the CI pipeline to use a feature flag seemed like a very high effort barrier. We also felt the scenario outline approach was marginally superior to copying and pasting the entire scenario.