Fix Orphan Document scenario for Document Collection
Document Collection Groups assume that documents added to the document groups always have a latest_edition attached to it since users are only allowed to search for and add published documents to the group. However, since there is behaviour to unpublish and delete the edition, the document is left orphaned, which is currently causing server errors for users who have previously added this document to their Document Collection Group.

Fix is to render the unavailable document so that the user would know to remove this orphaned document relationship from their document collection group.

Co-Authored-By: Nikin Nagewadia <nikin.nagewadia@digital.cabinet-office.gov.uk>