Handle validation errors
Attempting to set an email override field on a document that has already
been published will raise a validation error. Update the controller to handle
this exception. This error is not shown to the user as a flash, because in the next
commit, we make it impossible to access the document collection email subscription
edit forms for published document collections.