Add data migration to replace lead images which are SVGs
After a quick bit of investigation i can see that 5 pretty old draft
editions have had lead images created in the db which use SVGs.

This migration deletes any lead images which are SVGs and replaces them
with the latest non-svg image.

All of these editions only actually have 1 image so all that will happen
is they'll be removed as the lead image, but i've whacked the logic
in around updating the lead image in case anyones local data is in
a different state.