Add description to past prime ministers page
It was pointed out that the past prime ministers page was missing a
description. This means that only the title of the page shows up in the
search results.

The description has been hard-coded to match the page title, though
this content should probably be moved to a locale file or be editable
by publishers in the UI.

The content for the description also needs to be checked by a content
designer.