Manage image update and delete for promotional feature items
The previous behaviour manually deleted assets from asset-manager when
a new image is uploaded or a youtube video is uploaded instead. This behaviour
can also be achieved by introducing a carrierwave setting in the uploaders:
`remove_previously_stored_files_after_update` (see `FeaturedImageUploader`).

Nonetheless, this config option  is typically turned off for most document types, as to avoid
any 404 that may occur on published documents if their assets are missing. Whilst most documents
and their associations are republished, if anything goes wrong or the republishing takes a long
time to finish, the published pages will 404.

Keeping the assets ensure we avoid such issues at the cost of keeping overriden assets
in asset-manager.

We still delete corresponding assets when the `PromotionalFeatureItem` is destroyed.
The destroy on the parent classes causes the delete method in `asset_manager_storage`
to be called, which in turns enqueues the delete worker for asset-manager and then
deletes the assets in whitehall as well.