Fix file caching behaviour
The Carrierwave flow also gets triggered for the cached files, even when a file has been provided.
This means additional assets are created in `asset-manager` that will never be used.

In the previous legacy implementation, all the assets thus generated had the same
`legacy_url_path` and only the last ones were retrieved when querying the DB. As such,this posed
no other issues other than redundant data. Now that we are using asset IDs to communicate
with asset-manager, the last call on the `CreateAssetWorker` (which is the cache call) actually
overrides the previous assets in WH, meaning we do not see the expected image. In order to address
this, we filter out the `image_cache` param if an `image` one has been provided.