Remove graphviz_transitions dependency
This was added in 2c7c7df4ed, but I suspect nobody knows it's there or
how to use it.

The answer is:

```
rake transitions:draw CLASS=Edition
```

Which used to produce a nice PNG image (which I'll attach to the PR).

Unfortunately, this has been broken since rails 5 because of a breaking
change to the `.parameterize()` method - https://github.com/itkin/graphviz_transitions/pull/1

Less dependencies is better dependencies, so let's bin it.