Handle republishing related to assets for promotional feature item
The data related to promotional feature item is sent to publishing-api via the
`OrganisationPresenter`. Since the image is a required field as per the publishing api
schema, we are filtering out all promotional items with unhealthy assets, rather than
sending broken links (carrierwave magic produces a legacy-path-like url when assets are not
ready).

In order to make sure the correct urls are sent when they become available, we enqueue
an Organisation republish once the `CreateAssetWorker` finishes creating the assets.
The pattern we have used here is to define a method `republish_on_assets_ready` and call it,
if defined, on whatever model that has triggered the carrierwave flow and implicitly the
execution of the worker. Any model with a `mount_uploader` on it should also have a
`republish_on_assets_ready` method defined on it, to manage its republishing, as well
the republishing of its associations where required.