Use republishing method from Whitehall::PublishingApi instead of call publishworker directly
Republish method sets an update_type flag on the published content that has different side-effects than
the default flag. For example on regular publish, a 'major' value is set as the update_type and that
causes emails to be sent of the update. Republish uses 'republish' flag which means that a change note is not
required and there are no extra emails sent.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>