Handle republishing and filtering related to assets for Feature
The data related to feature is sent to publishing-api via the `OrganisationPresenter`.
We are filtering out all features with unhealthy assets, rather than sending broken
links (carrierwave magic produces a legacy-path-like url when assets are not ready).

In order to make sure the correct urls are sent when they become available, we enqueue
an Organisation republish once the `CreateAssetWorker` finishes creating the assets.
The pattern we have used here is to define a method `republish_on_assets_ready` and call it,
if defined, on whatever model that has triggered the carrierwave flow and implicitly the
execution of the worker. Any model with a `mount_uploader` on it should also have a
`republish_on_assets_ready` method defined on it, to manage its republishing, as well
the republishing of its associations where required.

For Feature, its republishing action is delegate to `FeatureList` which knows about the
featurable, which can be an `Organisation` or `WorldLocationNews`. Its associates Speeches
also get republished, as they may be using the Feature's image.

Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>
(cherry picked from commit ebffe1884db82a2e0a8786990045e3e6f244d84f)