Provide medium and high resolution images for features
In order to orchestrate a release in conjunction with `collections` application, url field has to be legacy url path temporarily.

Medium and High resolution images are required because `collections` app is expecting s465 and s712 versions of images. Previously the url of different versions of images were guessable from the original legacy path.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>
Co-authored-by: Tuomas Nylund <tuomas.nylund@digital.cabinet-office.gov.uk>