Ensure user gets readable errors when link validation errors occur
At the moment, users are unable to force schedule editions when
the DataHygiene::GovspeakLinkValidator throws errors for invalid links.

They are provided no information about which link is invalid or how
to fix it despite us having access to that information pretty readily.

This takes a first stab at resolving the issue. It is not a great
solution or implementation, but it works.

A far better solution would be to move the sidebar notices partial
into a view component and handle rendering the link and fix in the same
format as that of broken links.