Make the Edition#previous_edition method do what its name suggests it should
Before this commit, the Edition#previous_edition method was actually fetching the last post-publication edition for the document other than itself. That is not the behaviour consumers will expect based on the function name.

This commit changes the behaviour to fetch the previous post publication edition relative to this edition. Given that all tests pass when we provide a created_at date for the previous edition older than the current edition, this seems safe. Hopefully this will reduce confusion and potential misuse of the function in the future.