Update past prime ministers page when pm or ex-pm is updated
The past prime ministers page https://www.gov.uk/government/history/past-prime-ministers
has a list of all the PMs.

Currenty, the people edit page doesn't trigger an update to occur
for it. This can lead to the page getting out of sync as the persons
name is different to the name presented on the page.

It should also probably reublish any historical accounts but i'm
going to look into that next.