Remove all Admin Legacy preview functionality
Previews from Admin Legacy are no longer used and need to be removed.
There is functionality to enhanceYouTubeVideoLinks, but manually testing this, it doesn't seem to be replacing anything, could be due to how we are rendering the preview panel.
Either way, these files are not included in application.js so isn't included in the asset bundle to go to production anyways, hence currently doesn't seem to have any use.