Add UserOrderableClass concern for Orgs and MinisterialRoles
This adds a concern which includes a class method which handles reordering
records.

It applies to 2 classes. Organisations and Ministerial Roles.

These can be reordered in the Cabinet ministers order section of
the application.

Unfortunately we're going to need to retain update_column for this.
Now callbacks are handled asynchronously there's a legitimate chance that
the ordering could get out of sync if we call #update! on each ordering
being updated depending on which jobs get processed first/last.