Use UserOrderableClass#reorder! for take part pages
This follows the pattern established in this PR
https://github.com/alphagov/whitehall/pull/8557
for active record associations of adding a second reorder method with
doesn't skip callbacks and validations.

At the moment, the after_commit callbacks here are handlded syncronously
so the endpoint is really slow.

I'm going to add a seond commit following this that moves republishing
the get involved page out of a callback and into the controller actions
because it's simple and take part pages are isolated so there's no risk
of them being interacted with outside of this small section of Whitehall.