Introduce additional logic for checking that all asset variants are uploaded
Document types such as Person, Organisation, WorldwideOrganisation and PromotionalFeatureItem
have summary pages that may render faster than the assets can become available, as the redirection is
called straight after saving, and the image asset creation runs async in the background.

On creation, there are no pre-existing assets, so the previous checks suffice (i.e., mismatch between the number
of existing and expected assets). Nonetheless, on update, assets will already be in place from the previous image.
Since we update the image and assets in place (based on ID), and the assets get created asynchronously, it can happen
that the previous image's assets have not yet been written over so the old image may render. For example, when updating
a Person, the same FeaturedImageData ID is used for the update image. A number of 7 associated assets get generated on
create and only their `asset_manager_id` field gets updated.

The added checks in this commit ensure that the assets filename match the ones in the new image, so that we don't check against
stale assets.