Update flash notification message for person and edition images
Upon successful creation of a person or edition image, the page redirects to the show action
of the relevant controller, with a default success message. We have previously implemented a
PROCESSING label that is displayed when the image is still uploading. To give some context
to the user, the flash text has been updated to conditionally include some information about the
state of the image, i.e. if the image assets have not all finished processing.