Add active record extension that reorders association skipping validations
We have tonnes of duplication in our codebase around how we reorder a
collection of related objects for a model. For the most part we grab the
ordering params passed through by the Reorderable list Publishing
Component and use update_column to reorder the collection using the ordering
column.

Skipping validations here isn't ideal but sometimes it's required. This
commit handles updating the two collectons i've found that require callbacks
or validations to be skipped.

Topical event organisations have a db level unique constraint on the ordering index.

Topical event featurings republish the topical event each time a featuring
is updated via a callback.

Maybe that's fine and it can still use update, but if it has 100s of featurings
that's a lot of jobs called.