Update taggable branches
Update the branches shown to users in the email override taxonomy topic
select. This select is part of the Email Notifcations UI for document
collection pages. It will be used by HMRC to support a wider piece of
work to retire specialist topics. The list of branches has come from the
department.