Update PublishingApiDocumentRepublishingWorker to handle invalid drafts
We want to be able to republish documents an editions assoications are
updated via this worker. At the moment, we don't do this and it's often
not intuitive to the user at all that they must go to the edit edition
page and click save to send the changes downstream.

We're trying to ensure that when an assoication gets updated, these changes
are reflected downstream in the draft content store.

This updates the PublishingApiDocumentRepublishingWorker to not publish
a draft edition downstream if it is invalid.

This should largely open happen when an edition is carried over from a
post publication state to draft. At this point if the publisher doesn't
provide a change note or mark it as a minor change, they can go off
and for the most part edit associations, such as responses, document
collection groups etc.