Add procesing label for topical event
If not all assets are uploaded, the image load won't be sent to publishing api.

In order to make it known to the user that this might be the case, we display a
PROCESSING label when image assets are still uploading.

The label will stay on indefinitely if something is wrong with an asset upload.