Override corporate information page form template to change cancel link
The cancel routes helper is a very inelegant way to work around the fact that our edition template partials are not being used via composition.

I have therefore chosen to unDRY the corporate information page template so that it is not as tightly coupled to the basic edition template.

This enables us to override the cancel link path at the template level, where we know exactly what type of document we are dealing with, rather than having to re-evaluate what sort of document we have within a view helper.