Grafecully stop processing attachments when the related document has been destroyed
An issue in production showed that sometimes user might have uploaded a file and
for whatever reason, the create asset worker executes much later. So much so that
the user might have destroyed the actual original edition.

If this happens, the worker should stop processing the attachment instead of failing
and forever retrying.

Co-Authored-By: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>