Remove document summary components that use document_hub flag
We're no longer going with the previously designed approach so we should
remove the changes we made. We added a couple of view components for the
attachments and translations sections.

It might still be worth putting these into view components though because
the code is fairly unreadable and reasonably sparsely tested.