Ensure dashboard guidance links relate to header
Currently, the headings sit in one row and the guidance in a second below.
This causes screenreaders to read all the headings prior to hitting the
links which is confusing as users are not sure which link belongs to
which heading.

This resolves the issue by moving the links into the same row as the
headings.