Remove send call on Role to prevent injection attack
In the organisation roles controller #send is used directly on the
Role class.

This provides an attack vector for scripting. Particularly because it's
using #send in place of #public_send so it has access to the private
interface too.

There's some hidden complexity here because the type param is actually one
of 6 scopes called on the Role model.

This is a pretty rough and ready fix and there are probably better solutions.
I'd love to give this whole bit of the codebase a refactor as there's
a bunch of weird stuff going on, but for now i think this is probably
the simplest way to resolve the vunerability.

Ideally, i think the type query param would move to the path so we can
use constraints to match on type
https://guides.rubyonrails.org/routing.html#specifying-constraints

We could also potenitally add a separate reorder endpoint for each type
like we did for the CabinetMinistersController. But i think that's
probably a bit beyond the scope of this investigation & fix.

I'll be adding a card to the backlog to come back and look at this bit of
the codebase which i'll link in the PR.