Fix typo in partial name
This partial was renamed from `world_locations/list` to
`admin/world_location_news/list` when the WorldLocation model was split
into two parts.

Additionally, a route helper within the partial was renamed.

It appears the name of this partial was not updated here, and only just
became apparent when a somewhat related test was added part way through
development of the editionable worldwide organisations.