Add `Edition::Roles`
This will be used to allow roles to be assigned to editionable
documents.

The current usage will be for Editionable Worldwide Organisations, but
it is generic enough to be attached to any model by simply including
`include Edition::Roles`.

Nearly all the code has been copied verbatim from
`Edition::WorldLocations`, including the tests.