Allow roles to be associated with Editionable Worldwide Organisations
This is a change from the current behaviour, as the worldwide
organisations are currently assigned on the role's edit page.

By moving this to be the other way round (i.e. the edit is made on the
worldwide organisation itself), the option is more obvious to the user
and allows the change to be made in advance (e.g. as part of scheduled
publishing).