Fix Document Collections to only search for published editions
When we introduced pagination, we inadvertently removed the scope of published editions.
In addition to putting the scope back to how it was, we also removed the filtering since we didn't need to do any other filtering other than search via title. This should simplify both the code and performance.
Hopefully the scoping will also be helpful for search performance in production as well.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>
Co-authored-by: Ryan Brown <ryan.brown@digital.cabinet-office.gov.uk>