Add an index to the asset manager ID column on the assets table
The lack of an index was causing lock timeout errors occasionally during concurrent delete operations.

The asset deletion worker runs a query to delete all assets matching a particular ID in asset manager. For delete queries, InnoDB will issue a lock for each record that is scanned, regardless of whether or not that record will subsequently be deleted. Consequently, adding an index to reduce the number of records scanned will reduce the risk of locks caused by concurrent deletes.