Add view to manage editionable social media links
This will allow us to manage social media links that are associated with
editions.

At the moment, this lists the existing accounts (with no action links)
and a dummy link to add a new social media account.

The actions will be populated in later commits.

This form is generic so it can be used for any editionable documents
that are associated with social media links, not just worldwide
organisations.