Update `SocialMediaAccount` factory
We are currently using this factory in a number of tests and checking
the title, but never actually setting a title.

Updating this so a title is set and so the attributes are unique for
each instance.

Also updates two tests that assume the title will always be `nil` and
fixes a faulty test that is checking the URL of the wrong instance (but
was never previously an issue as the URL was always the same for each
instance).