Make the worldwide offices work for editionable worldwide organisations
This modifies the existing worldwide offices controller and model so that they
are compatible with editionable worldwide organisations as well as legacy
worldwide organisations.

Note: When the editionable worldwide organisations feature flag is enabled,
performing actions related to the worldwide offices of legacy worldwide
organisations will not function (and vice versa). This should be okay as we
don't want these features running side-by-side. The next step will be to hide
the legacy actions when the flag is enabled for an extra level of safety.