Move delete button for social media accounts
Moves the delete button for social media accounts that are associated
with editions to be alongside the account, rather than next to each
translation.

This is because we want to delete the social media account overall, not
just a specific translation.

Matches the location of the button for social media accounts that are
not associated with editions.