Raise if the person presenter attempts to present carrierwave temp URL to Publishing API
This is a temporary measure to make sure we get a good error report when this happens in production. Currently the error is difficult to diagnose because we don't have a stack trace, as we don't know that there's a problem until the user reports that the image is not being rendered on the frontend.