Ensure GDS Admin can search for limited access documents
We're doing some work to allow GDS Admins to remove limited access/and
or update the organisations for one.

In order to do this GDS Admins will need to be able to search for
the document before being able to make the updates.

This updates the EditionFilter to permit access limited documents
for GDS Admins.