Add searchable filter to roles table
This adds a searchable filter which allows users to search on the name
of the role, the current appointee etc.

We've been asked to do this as part of our work on gift week to make
our Content Designers lives easier.

We've added some text when there's no appointment so that users can
filter for vacant appointments and clean them up, which is a common
task.