Update sentry-rails to 5.16.1
A bad release of sentry-sidekiq yesterday
(https://github.com/getsentry/sentry-ruby/issues/2221) caused sentry-rails to
be downgraded to 5.3.1 rather than updated.

As a result, we are seeing some odd behaviour where instead of being shown
error pages specific to the error, we are always shown the production server
error page. This is present in production, but also extends to the development
environment where we no longer see the developer exception pages.

This brings sentry-rails in line with sentry-ruby and fixes the issue.