Inform the user that an invalid filter has not been applied
There was no validation on last updated from and to dates filters. As a result,
the user was redirected to session stored filters or default filter (organisation and active state),
without being told so.

Adding a validation check for dates and alert to inform the user that the filters are invalid.