Ensure Republishing worker can re-unpublish HTML attachments for unpublished editions successfully
When the republishing worker is passed an unpublished document, it has to republish the document in case there have been code changes that would affect the display of the unpublishing message (e.g. govspeak changes), and then unpublish it again. It also has to do this for the unpublished document's HTML attachments.

It might be that we can remove this behaviour, because the route for an unpublished document should either redirect or show a generic "gone" message, but I'm not confident enough to risk it yet.

In any case, currently the unpublishing of the HTML attachments fails. Publishing API returns a 404 response. This is because if the "allow_draft" parameter is used, which we were using, Publishing API tries to find a document from the draft stack to unpublish. However, there isn't one there, because the republishing worker has just published the attachment (to potentially update the Govspeak etc.).

We have therefore fixed this issue by switching the "allow_draft" parameter to false for the unpublishing request.

We need to test the  behaviour of republishing worker for unpublished documents with drafts using Webmock. We need to use Webmock because mocking the republishing worker dependencies is obscuring an issue which causes HTML attachments to published inadvertently whilst republishing an unpublished document