Use edition links for worldwide organisations
We are now using edition links for editionable worldwide organisations.
This makes the same change for non-editionable worldwide organisations.

There will be no user impact as a result of this change, but it will
allow us to compare the output of the two presenters to make sure they
are identical when migrating data from the non-editionable worldwide
organisation to be editionable.

The existing links must be set as an empty array to clear them (as an
nil value does not unset the existing values). Another PR will be raised
to remove the non-edition links from the presenter once all worldwide
organisations have been republished.

Depends on https://github.com/alphagov/publishing-api/pull/2601.